// stade-agadir/src/utils/interfaces/map/EStadiumIds.ts
export enum EStadiumIds {
    TRIBUNE = "tribune",
    TRIBUNE_LINE = "tribune-line",
    SEAT = "seat",
    SEAT_LINE = "seat-line",
  
    FIELD = "field",
    TRIBUNE_SYMBOL = "symbol-layer",
    STADIUM_OUTLINE = "stadium-outline",
    STADIUM_BASE = "stadium-base",
    BG_STADIUM = "bg-stadium",
    BG_STADIUM_LINE = "bg-stadium-line",
  }